@@ -1,3 +1,10 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2022-04-27 16:50:25
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2022-07-05 14:14:04
+ * @Description: 宽高等适配
+ */
 import 'dart:ui';
 import 'dart:math'; // min
 import 'package:flutter/material.dart';
@@ -6,19 +13,42 @@ class AdaptCJHelper {
   static bool _hasInit = false;
 
   static Size uiSize = const Size(375, 812);
-  static double _screenWidth;
-  static double _screenHeight;
-  static double _scaleWidth;
-  static double _scaleHeight;
+  static late double _screenWidth;
+  static late double _screenHeight;
+  static late double _stautsBarHeight;
+  static late double _screenBottomHeight;
+  static late double _scaleWidth;
+  static late double _scaleHeight;
+
   static init() {
     MediaQueryData mediaQuery = MediaQueryData.fromWindow(window);
     _screenWidth = mediaQuery.size.width;
     _screenHeight = mediaQuery.size.height;
+    _stautsBarHeight = mediaQuery.padding.top;
+
+    ///TODO:疑难杂症:这里为什么有时候会等于0，常见于点击搜索后，弹窗键盘,回来发现
+    _screenBottomHeight = mediaQuery.padding.bottom;
 
     _scaleWidth = _screenWidth / uiSize.width;
     _scaleHeight = _screenHeight / uiSize.height;
 
     _hasInit = true;
+  }
+
+  static double get stautsBarHeight {
+    if (_hasInit != true) {
+      AdaptCJHelper.init();
+    }
+
+    return _stautsBarHeight;
+  }
+
+  static double get screenBottomHeight {
+    if (_hasInit != true) {
+      AdaptCJHelper.init();
+    }
+
+    return _screenBottomHeight;
   }
 
   /// 实际尺寸与UI设计的比例
