@@ -1,7 +1,14 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2022-04-15 22:08:25
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2022-07-07 13:44:02
+ * @Description: 
+ */
 class IndexPath {
   final int section;
   final int row;
-  IndexPath({this.section, this.row});
+  IndexPath({required this.section, required this.row});
   @override
   String toString() {
     return 'section_${section}_row_$row';
@@ -14,7 +21,7 @@ class IndexPath {
     if (other.runtimeType != IndexPath) {
       return false;
     }
-    IndexPath otherIndexPath = other;
+    IndexPath otherIndexPath = other as IndexPath;
     return section == otherIndexPath.section && row == otherIndexPath.row;
   }
 }
