@@ -2,16 +2,15 @@
  * @Author: dvlproad
  * @Date: 2022-04-12 23:04:04
  * @LastEditors: dvlproad
- * @LastEditTime: 2022-04-15 18:28:19
+ * @LastEditTime: 2022-05-13 19:54:45
  * @Description: 基础UI库
  */
 library flutter_baseui_kit;
 
-export './base-uikit/bg_border_widget.dart';
-export './base-uikit/button/imagebutton.dart';
 // text
 export './cq-uikit/text/tolerant_text.dart';
 
+// button
 export './cq-uikit/button/themtextbutton.dart';
 export './cq-uikit/button/statetextbutton.dart';
 export './cq-uikit/button/buttontheme.dart'
@@ -20,13 +19,21 @@ export './cq-uikit/button/buttontheme.dart'
         ButtonRegularTextStyle,
         ButtonBoldTextStyle,
         ButtonMediumTextStyle;
-
+// image button
 export './cq-uikit/button/imagebutton.dart';
 
+// inkWelll
+export './cq-uikit/button/ink_well.dart';
+
+// textinput
+export './cq-uikit/textfield/prefixIcon_textField.dart';
+export './cq-uikit/textfield/prefixText_textField.dart';
+
+// cell
 export './cq-uikit/cell/cell.dart';
 
+// tableview
 export './cq-uikit/tableview/section_table_view_method2.dart';
-export './cq-uikit/tableview/index_path.dart';
 
 // toolbar
 export './cq-uikit/toolbar/toolbar.dart';
@@ -39,3 +46,5 @@ export './cq-uikit/absorbBottom/absorb_bottom_widget.dart';
 
 // 点击优化
 export './opt/click_optimization.dart';
+
+export './cq-uikit/text/skeleton.dart.dart';
